requirements.txt
fastapi
uvicorn
aiosmtplib
python-dotenv
jinja2, 
python-multipart